chore(types): Move ApiResponse typing from ApplicationState to seperate file
